# perf: remediate non-sargable leaderboard queries