--- a/app/Helpers/database/leaderboard.php
+++ b/app/Helpers/database/leaderboard.php
@@ -206,16 +206,16 @@ function getLeaderboardsForGame(int $gameID, ?array &$dataOut, ?string $localUse
               FROM LeaderboardDef AS lbd
               LEFT JOIN (
                   SELECT lbd.ID as LeaderboardID,
-                  CASE WHEN !lbd.LowerIsBetter THEN MAX(le2.Score) ELSE MIN(le2.Score) END AS Score
+                  CASE WHEN lbd.LowerIsBetter = 0 THEN MAX(le2.Score) ELSE MIN(le2.Score) END AS Score
                   FROM LeaderboardDef AS lbd
                   LEFT JOIN LeaderboardEntry AS le2 ON lbd.ID = le2.LeaderboardID
                   LEFT JOIN UserAccounts AS ua ON ua.ID = le2.UserID
-                  WHERE !ua.Untracked AND lbd.GameID = $gameID
+                  WHERE ua.Untracked = 0 AND lbd.GameID = $gameID
                   GROUP BY lbd.ID
               ) AS BestEntries ON BestEntries.LeaderboardID = lbd.ID
               LEFT JOIN LeaderboardEntry AS le2 ON le2.LeaderboardID = lbd.ID AND le2.Score = BestEntries.Score
               LEFT JOIN UserAccounts ua ON le2.UserID = ua.ID
-              WHERE lbd.GameID = $gameID AND (ua.User IS NULL OR !ua.Untracked)
+              WHERE lbd.GameID = $gameID AND (ua.User IS NULL OR ua.Untracked = 0)
               ORDER BY DisplayOrder ASC, LeaderboardID, DateSubmitted ASC";
 
     $dataOut = [];
@@ -258,8 +258,8 @@ function GetLeaderboardEntriesDataJSON(int $lbID, string $user, int $numToFetch,
               LEFT JOIN LeaderboardDef AS lbd ON lbd.ID = le.LeaderboardID
               WHERE le.LeaderboardID = $lbID AND $friendQuery
               ORDER BY
-              CASE WHEN !lbd.LowerIsBetter THEN Score END DESC,
-              CASE WHEN lbd.LowerIsBetter THEN Score END ASC, DateSubmitted ASC
+              CASE WHEN lbd.LowerIsBetter = 0 THEN Score END DESC,
+              CASE WHEN lbd.LowerIsBetter = 1 THEN Score END ASC, DateSubmitted ASC
               LIMIT $offset, $numToFetch ";
 
     $dbResult = s_mysql_query($query);
@@ -308,7 +308,7 @@ function GetLeaderboardData(
           SELECT COUNT(UserID)
           FROM LeaderboardEntry AS le
           LEFT JOIN UserAccounts AS ua ON ua.ID = le.UserID
-          WHERE !ua.Untracked AND le.LeaderboardID = $lbID
+          WHERE ua.Untracked = 0 AND le.LeaderboardID = $lbID
         ) AS TotalEntries
       FROM LeaderboardDef AS ld
       LEFT JOIN GameData AS gd ON gd.ID = ld.GameID
@@ -348,18 +348,18 @@ function GetLeaderboardData(
 
         // Now get entries:
         $query = "SELECT ua.User, le.Score, le.DateSubmitted,
-                  CASE WHEN !lbd.LowerIsBetter
+                  CASE WHEN lbd.LowerIsBetter = 0
                   THEN RANK() OVER(ORDER BY le.Score DESC)
                   ELSE RANK() OVER(ORDER BY le.Score ASC) END AS UserRank,
-                  CASE WHEN !lbd.LowerIsBetter
+                  CASE WHEN lbd.LowerIsBetter = 0
                   THEN ROW_NUMBER() OVER(ORDER BY le.Score DESC, le.DateSubmitted ASC)
                   ELSE ROW_NUMBER() OVER(ORDER BY le.Score ASC, le.DateSubmitted ASC) END AS UserIndex
                   FROM LeaderboardEntry AS le
                   LEFT JOIN UserAccounts AS ua ON ua.ID = le.UserID
                   LEFT JOIN LeaderboardDef AS lbd ON lbd.ID = le.LeaderboardID
-                  WHERE (!ua.Untracked || ua.User = '$user' ) AND le.LeaderboardID = $lbID
+                  WHERE (ua.Untracked = 0 || ua.User = '$user' ) AND le.LeaderboardID = $lbID
                   ORDER BY
-                  CASE WHEN !lbd.LowerIsBetter THEN Score END DESC,
+                  CASE WHEN lbd.LowerIsBetter = 0 THEN Score END DESC,
                   CASE WHEN lbd.LowerIsBetter THEN Score END ASC, DateSubmitted ASC
                   LIMIT $offset, $numToFetch ";
 
@@ -392,16 +392,16 @@ function GetLeaderboardData(
                 // Go find user's score in this table, if it exists!
                 $query = "SELECT User, Score, DateSubmitted, UserRank, UserIndex FROM
                          (SELECT ua.User, le.Score, le.DateSubmitted,
-                          CASE WHEN !lbd.LowerIsBetter
+                          CASE WHEN lbd.LowerIsBetter = 0
                           THEN RANK() OVER(ORDER BY le.Score DESC)
                           ELSE RANK() OVER(ORDER BY le.Score ASC) END AS UserRank,
-                          CASE WHEN !lbd.LowerIsBetter
+                          CASE WHEN lbd.LowerIsBetter = 0
                           THEN ROW_NUMBER() OVER(ORDER BY le.Score DESC, le.DateSubmitted ASC)
                           ELSE ROW_NUMBER() OVER(ORDER BY le.Score ASC, le.DateSubmitted ASC) END AS UserIndex
                           FROM LeaderboardEntry AS le
                           LEFT JOIN UserAccounts AS ua ON ua.ID = le.UserID
                           LEFT JOIN LeaderboardDef AS lbd ON lbd.ID = le.LeaderboardID
-                          WHERE !ua.Untracked AND le.LeaderboardID = $lbID) InnerTable
+                          WHERE ua.Untracked = 0 AND le.LeaderboardID = $lbID) InnerTable
                           WHERE InnerTable.User = '$user'";
 
                 $dbResult = s_mysql_query($query);
@@ -512,13 +512,13 @@ function getLeaderboardUserPosition(int $lbID, string $user, ?int &$lbPosition):
 
     $query = "SELECT UserIndex FROM
                          (SELECT ua.User, le.Score, le.DateSubmitted,
-                          CASE WHEN !lbd.LowerIsBetter
+                          CASE WHEN lbd.LowerIsBetter = 0
                           THEN ROW_NUMBER() OVER(ORDER BY le.Score DESC, le.DateSubmitted ASC)
                           ELSE ROW_NUMBER() OVER(ORDER BY le.Score ASC, le.DateSubmitted ASC) END AS UserIndex
                           FROM LeaderboardEntry AS le
                           LEFT JOIN UserAccounts AS ua ON ua.ID = le.UserID
                           LEFT JOIN LeaderboardDef AS lbd ON lbd.ID = le.LeaderboardID
-                          WHERE !ua.Untracked AND le.LeaderboardID = $lbID) InnerTable
+                          WHERE ua.Untracked = 0 AND le.LeaderboardID = $lbID) InnerTable
                           WHERE InnerTable.User = '$user'";
 
     $dbResult = s_mysql_query($query);