There are several leaderboard queries, including one that executes on game pages, that are non-sargable and not able to use indexes properly.

**Benchmarks**
These samples were collected on game ID 1446 with the local MySQL cache disabled.

|         | Before | After |
|---------|--------|-------|
| 1       | 499ms  | 383ms |
| 2       | 335ms  | 280ms |
| 3       | 444ms  | 260ms |
| 4       | 334ms  | 323ms |
| 5       | 339ms  | 270ms |
| 6       | 488ms  | 265ms |
| 7       | 310ms  | 334ms |
| 8       | 303ms  | 262ms |
| 9       | 292ms  | 260ms |
| Average | 371ms  | 293ms |

Speedup Factor: ~1.26x